refactor: css fonts changed + frontend project structure reorganized
