--- front-end/src/user/UserProfileButton.css
@@ -1,13 +1,0 @@
-
-.user-profile-button{
-    
-}
-.profile-icon{
-    width: 25%;
-    height: 25%;
-    left:0;
-}
-.auth-button{
-    background-color: lightblue;
-    display:inline;
-}
