+++ front-end/src/user/styles/UserProfileButton.css
@@ -1,0 +1,24 @@
+
+.profile-text{
+    font-family: Superclarendon, 'Bookman Old Style', 'URW Bookman', 'URW Bookman L', 'Georgia Pro', Georgia, serif;
+    font-weight: 500;
+}
+.profile-icon{
+    width: 15%;
+    height: 15%;
+    left:0;
+}
+.auth-button{
+    display: flex;
+    align-items: center; 
+    gap: 10px; 
+    border: 1px solid black;
+    padding: 10px;
+    cursor: pointer;
+    border-radius:4px;
+}
+
+.auth-button:hover{
+    background: #473dff;
+    transform: scale(1.05);
+}
