+++ src/main/java/com/marsannar2/personalfinance/user/AppUserController.java
@@ -1,11 +1,7 @@
 package com.marsannar2.personalfinance.user;
 
-import java.util.stream.Collectors;
-
-import org.apache.catalina.connector.Response;
-import org.apache.tomcat.websocket.AuthenticationException;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -18,18 +14,11 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 
 import com.marsannar2.personalfinance.dto.user.LoginRequest;
 import com.marsannar2.personalfinance.dto.user.SignInRequest;
 import com.marsannar2.personalfinance.utils.response.MessageResponse;
 
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
 @RestController
